Added price value to productSaveRequest

+++ src/controllers/commands/models/productModel.ts
@@ -27,7 +27,7 @@ ProductModel.init(
 		},
 		price: <ModelAttributeColumnOptions>{
 			field: ProductFieldName.PRICE,
-			type: DataTypes.BIGINT,
+			type: DataTypes.DECIMAL(10, 2),
 			allowNull: true
 		},
 		createdOn: <ModelAttributeColumnOptions>{
